requirements.txt
sounddevice
scipy
numpy
openai
whisper
python-dotenv
selenium
